move entity type stuff to helper
